[XLA:GPU] Speed up layout assignment.
In some cases the layout assignment pass takes too much time. That happens when we migrate the long list of users of one instruction to another one one by one. In this specific case we want to move all of them to a new producer except the newly inserted follow instruction.

Let's do that by swapping the entire users_ property between the old producer and the new producer.

PiperOrigin-RevId: 684806386